Move generic commons to upper package
