+++ captum/attr/_core/feature_permutation.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, Tuple, Union
 import torch
 from torch import Tensor
 
-from .._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
+from ..._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 from .feature_ablation import FeatureAblation
