+++ captum/attr/_core/guided_backprop_deconvnet.py
@@ -8,10 +8,11 @@ from torch import Tensor
 from torch.nn import Module
 from torch.utils.hooks import RemovableHandle
 
+from ..._utils.common import _format_input, _is_tuple
+from ..._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 from .._utils.attribution import GradientAttribution
-from .._utils.common import _format_attributions, _format_input, _is_tuple
+from .._utils.common import _format_attributions
 from .._utils.gradient import apply_gradient_requirements, undo_gradient_requirements
-from .._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 
 
 class ModifiedReluGradientAttribution(GradientAttribution):
