+++ captum/attr/_core/input_x_gradient.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python3
 from typing import Any, Callable
 
+from ..._utils.common import _format_input, _is_tuple
+from ..._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 from .._utils.attribution import GradientAttribution
-from .._utils.common import _format_attributions, _format_input, _is_tuple
+from .._utils.common import _format_attributions
 from .._utils.gradient import apply_gradient_requirements, undo_gradient_requirements
-from .._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 
 
 class InputXGradient(GradientAttribution):
