+++ captum/attr/_core/layer/internal_influence.py
@@ -5,20 +5,22 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
+from ...._utils.common import (
+    _expand_additional_forward_args,
+    _expand_target,
+    _format_additional_forward_args,
+)
+from ...._utils.typing import BaselineType, TargetType
 from ..._utils.approximation_methods import approximation_parameters
 from ..._utils.attribution import GradientAttribution, LayerAttribution
 from ..._utils.batching import _batched_operator
 from ..._utils.common import (
-    _expand_additional_forward_args,
-    _expand_target,
-    _format_additional_forward_args,
     _format_attributions,
     _format_input_baseline,
     _reshape_and_sum,
     _validate_input,
 )
 from ..._utils.gradient import compute_layer_gradients_and_eval
-from ..._utils.typing import BaselineType, TargetType
 
 
 class InternalInfluence(LayerAttribution, GradientAttribution):
