+++ captum/attr/_core/neuron/neuron_feature_ablation.py
@@ -4,10 +4,10 @@ from typing import Any, Callable, List, Tuple, Union
 import torch
 from torch.nn import Module
 
+from ...._utils.common import _verify_select_column
+from ...._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 from ..._utils.attribution import NeuronAttribution, PerturbationAttribution
-from ..._utils.common import _verify_select_column
 from ..._utils.gradient import _forward_layer_eval
-from ..._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 from ..feature_ablation import FeatureAblation
