+++ captum/attr/_core/neuron/neuron_gradient.py
@@ -3,19 +3,15 @@ from typing import Any, Callable, List, Tuple, Union
 
 from torch.nn import Module
 
+from ...._utils.common import _format_additional_forward_args, _format_input, _is_tuple
+from ...._utils.typing import TensorOrTupleOfTensorsGeneric
 from ..._utils.attribution import GradientAttribution, NeuronAttribution
-from ..._utils.common import (
-    _format_additional_forward_args,
-    _format_attributions,
-    _format_input,
-    _is_tuple,
-)
+from ..._utils.common import _format_attributions
 from ..._utils.gradient import (
     _forward_layer_eval_with_neuron_grads,
     apply_gradient_requirements,
     undo_gradient_requirements,
 )
-from ..._utils.typing import TensorOrTupleOfTensorsGeneric
 
 
 class NeuronGradient(NeuronAttribution, GradientAttribution):
