+++ captum/attr/_core/neuron/neuron_integrated_gradients.py
@@ -4,9 +4,9 @@ from typing import Any, Callable, List, Tuple, Union
 from torch import Tensor
 from torch.nn import Module
 
+from ...._utils.typing import TensorOrTupleOfTensorsGeneric
 from ..._utils.attribution import GradientAttribution, NeuronAttribution
 from ..._utils.gradient import construct_neuron_grad_fn
-from ..._utils.typing import TensorOrTupleOfTensorsGeneric
 from ..integrated_gradients import IntegratedGradients
