+++ captum/attr/_core/saliency.py
@@ -4,12 +4,11 @@ from typing import Any, Callable
 
 import torch
 
-from captum.attr._utils.typing import TensorOrTupleOfTensorsGeneric
-
+from ..._utils.common import _format_input, _is_tuple
+from ..._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 from .._utils.attribution import GradientAttribution
-from .._utils.common import _format_attributions, _format_input, _is_tuple
+from .._utils.common import _format_attributions
 from .._utils.gradient import apply_gradient_requirements, undo_gradient_requirements
-from .._utils.typing import TargetType
 
 
 class Saliency(GradientAttribution):
