+++ captum/attr/_utils/batching.py
@@ -5,8 +5,8 @@ from typing import Any, Callable, Dict, Iterator, List, Tuple, Union
 import torch
 from torch import Tensor, device
 
-from .common import _format_additional_forward_args, _format_input
-from .typing import (
+from ..._utils.common import _format_additional_forward_args, _format_input
+from ..._utils.typing import (
     TargetType,
     TensorOrTupleOfTensorsGeneric,
     TupleOrTensorOrBoolGeneric,
