+++ tests/attr/layer/test_internal_influence.py
@@ -6,8 +6,8 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
+from captum._utils.typing import BaselineType
 from captum.attr._core.layer.internal_influence import InternalInfluence
-from captum.attr._utils.typing import BaselineType
 
 from ..helpers.basic_models import (
     BasicModel_MultiLayer,
@@ -169,7 +169,7 @@ class Test(BaseTest):
                 attribute_to_layer_input=attribute_to_layer_input,
             )
             assertTensorTuplesAlmostEqual(
-                self, attributions, expected_activation, delta=0.01, mode="max",
+                self, attributions, expected_activation, delta=0.01, mode="max"
             )
