+++ tests/attr/layer/test_layer_ablation.py
@@ -7,8 +7,8 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
+from captum._utils.typing import BaselineType
 from captum.attr._core.layer.layer_feature_ablation import LayerFeatureAblation
-from captum.attr._utils.typing import BaselineType
 
 from ..helpers.basic_models import (
     BasicModel_ConvNet_One_Conv,
