+++ tests/attr/layer/test_layer_gradient_shap.py
@@ -5,9 +5,9 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
+from captum._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.gradient_shap import GradientShap
 from captum.attr._core.layer.layer_gradient_shap import LayerGradientShap
-from captum.attr._utils.typing import TargetType, TensorOrTupleOfTensorsGeneric
 
 from ..helpers.basic_models import (
     BasicModel_MultiLayer,
