+++ tests/attr/neuron/test_neuron_ablation.py
@@ -7,8 +7,8 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
+from captum._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.neuron.neuron_feature_ablation import NeuronFeatureAblation
-from captum.attr._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 
 from ..helpers.basic_models import (
     BasicModel_ConvNet_One_Conv,
