+++ tests/attr/test_feature_ablation.py
@@ -6,12 +6,8 @@ from typing import Any, Callable, List, Tuple, Union
 import torch
 from torch import Tensor
 
+from captum._utils.typing import BaselineType, TargetType, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.feature_ablation import FeatureAblation
-from captum.attr._utils.typing import (
-    BaselineType,
-    TargetType,
-    TensorOrTupleOfTensorsGeneric,
-)
 
 from .helpers.basic_models import (
     BasicModel,
@@ -217,13 +213,8 @@ class Test(BaseTest):
         model = BasicModelWithSparseInputs()
         inp1 = torch.tensor([[1.0, -2.0, 3.0], [2.0, -1.0, 3.0]])
         inp2 = torch.tensor([])
-        exp: Tuple[List[List[float]], ...] = (
-            [[9.0, -3.0, 12.0]],
-            [[]],
-        )
-        self._ablation_test_assert(
-            model, (inp1, inp2), exp, target=None,
-        )
+        exp: Tuple[List[List[float]], ...] = ([[9.0, -3.0, 12.0]], [[]])
+        self._ablation_test_assert(model, (inp1, inp2), exp, target=None)
 
     def test_sparse_features(self) -> None:
         model = BasicModelWithSparseInputs()
@@ -231,7 +222,7 @@ class Test(BaseTest):
         # Length of sparse index list may not match # of examples
         inp2 = torch.tensor([1, 7, 2, 4, 5, 3, 6])
         self._ablation_test_assert(
-            model, (inp1, inp2), ([[9.0, -3.0, 12.0]], [[2.0]],), target=None,
+            model, (inp1, inp2), ([[9.0, -3.0, 12.0]], [[2.0]]), target=None
         )
 
     def test_single_ablation_batch_scalar_float(self) -> None:
