+++ tests/attr/test_guided_backprop.py
@@ -6,11 +6,11 @@ from typing import Any, List, Tuple, Union
 import torch
 from torch.nn import Module
 
+from captum._utils.typing import TensorOrTupleOfTensorsGeneric
 from captum.attr._core.guided_backprop_deconvnet import GuidedBackprop
 from captum.attr._core.neuron.neuron_guided_backprop_deconvnet import (
     NeuronGuidedBackprop,
 )
-from captum.attr._utils.typing import TensorOrTupleOfTensorsGeneric
 
 from .helpers.basic_models import BasicModel_ConvNet_One_Conv
 from .helpers.utils import BaseTest, assertTensorAlmostEqual
