+++ tests/attr/test_guided_grad_cam.py
@@ -6,8 +6,8 @@ from typing import Any
 import torch
 from torch.nn import Module
 
+from captum._utils.typing import TensorOrTupleOfTensorsGeneric
 from captum.attr._core.guided_grad_cam import GuidedGradCam
-from captum.attr._utils.typing import TensorOrTupleOfTensorsGeneric
 
 from .helpers.basic_models import BasicModel_ConvNet_One_Conv
 from .helpers.utils import BaseTest, assertTensorAlmostEqual
