+++ tests/attr/test_integrated_gradients_basic.py
@@ -6,14 +6,10 @@ from typing import Any, Tuple, Union, cast
 import torch
 from torch.nn import Module
 
+from captum._utils.typing import BaselineType, Tensor, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.integrated_gradients import IntegratedGradients
 from captum.attr._core.noise_tunnel import NoiseTunnel
 from captum.attr._utils.common import _tensorize_baseline, _zeros
-from captum.attr._utils.typing import (
-    BaselineType,
-    Tensor,
-    TensorOrTupleOfTensorsGeneric,
-)
 
 from .helpers.basic_models import (
     BasicModel,
