+++ tests/attr/test_integrated_gradients_classification.py
@@ -5,9 +5,9 @@ import unittest
 import torch
 from torch.nn import Module
 
+from captum._utils.typing import BaselineType, Tensor
 from captum.attr._core.integrated_gradients import IntegratedGradients
 from captum.attr._core.noise_tunnel import NoiseTunnel
-from captum.attr._utils.typing import BaselineType, Tensor
 
 from .helpers.classification_models import SigmoidModel, SoftmaxModel
 from .helpers.utils import BaseTest, assertTensorAlmostEqual
